+++ common/src/core/dataprovider/media_streamdataprovider.cpp
@@ -138,7 +138,7 @@ bool QnAbstractMediaStreamDataProvider::afterGetData(const QnAbstractDataPacketP
         data->dataProvider = this;
 
     if (videoData)
-        m_stat[videoData->channelNumber].onData(data->dataSize());
+        m_stat[videoData->channelNumber].onData(static_cast<unsigned int>(data->dataSize()));
 
     return true;
